Добавлено бронирование и возможность оставить отзыв

+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImpl.java
@@ -1,71 +1,77 @@
 package ru.practicum.shareit.user.service;
 
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.practicum.shareit.exception.ConflictException;
 import ru.practicum.shareit.user.UserMapper;
 import ru.practicum.shareit.user.UserRepository;
 import ru.practicum.shareit.user.dto.UserDto;
 import ru.practicum.shareit.user.model.User;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
+    private final UserMapper userMapper;
 
     @Override
     public UserDto createUser(UserDto userDto) {
-        userRepository.findByEmail(userDto.getEmail())
-                .ifPresent(user -> {
-                    throw new IllegalArgumentException("Пользователь с таким email уже существует.");
-                });
+        if (userRepository.findByEmail(userDto.getEmail()).isPresent()) {
+            throw new ConflictException("Пользователь с email " + userDto.getEmail() + " уже существует");
+        }
 
-        User user = UserMapper.toUser(userDto);
+        User user = userMapper.toEntity(userDto);
         User savedUser = userRepository.save(user);
-        return UserMapper.toUserDto(savedUser);
+        return userMapper.toDto(savedUser);
     }
 
     @Override
     public UserDto updateUser(Long userId, UserDto userDto) {
-        User existingUser = userRepository.findById(userId)
-                .orElseThrow(() -> new NoSuchElementException("Пользователь не найден."));
+        User existingUser = getUserEntityById(userId);
 
         if (userDto.getName() != null) {
             existingUser.setName(userDto.getName());
         }
         if (userDto.getEmail() != null && !userDto.getEmail().equals(existingUser.getEmail())) {
-            userRepository.findByEmail(userDto.getEmail())
-                    .ifPresent(user -> {
-                        if (!user.getId().equals(userId)) {
-                            throw new IllegalArgumentException("Пользователь с таким email уже существует.");
-                        }
-                    });
+            userRepository.findByEmail(userDto.getEmail()).ifPresent(found -> {
+                if (!found.getId().equals(userId)) {
+                    throw new ConflictException("Email уже занят");
+                }
+            });
             existingUser.setEmail(userDto.getEmail());
         }
 
         User updatedUser = userRepository.save(existingUser);
-        return UserMapper.toUserDto(updatedUser);
+        return userMapper.toDto(updatedUser);
     }
 
     @Override
     public UserDto getUserById(Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new NoSuchElementException("Пользователь не найден."));
-        return UserMapper.toUserDto(user);
+        User user = getUserEntityById(userId);
+        return userMapper.toDto(user);
     }
 
     @Override
     public List<UserDto> getAllUsers() {
         return userRepository.findAll().stream()
-                .map(UserMapper::toUserDto)
+                .map(userMapper::toDto)
                 .collect(Collectors.toList());
     }
 
     @Override
     public void deleteUser(Long userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new EntityNotFoundException("Пользователь с id: " + userId + " не найден");
+        }
         userRepository.deleteById(userId);
+    }
+
+    private User getUserEntityById(Long userId) {
+        return userRepository.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException("Пользователь с id: " + userId + " не найден"));
     }
 }
